added_game_logic

+++ app/src/main/java/com/example/composition/domain/entity/GameSettings.kt
@@ -7,5 +7,5 @@ data class GameSettings(
     val maxSumValue: Int,
     val minCountOfRightAnswers: Int,
     val minPercentOfRightAnswers: Int,
-    val gameTimeInSeconds: Int
+    val gameTimeInSeconds: Long
 ) : Parcelable
